Create another main
