+++ ble_serial/__init__.py
@@ -1,0 +1,1 @@
+from ble_serial.main import Main
